relocate fallback function to livechat file
